+++ src/EditorFeatures/Core/TextDiffing/EditorTextDifferencingService.cs
@@ -12,6 +12,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.CodeAnalysis.Editor.Shared.Extensions;
 using Microsoft.CodeAnalysis.Host.Mef;
+using Microsoft.CodeAnalysis.Preview;
 using Microsoft.CodeAnalysis.Text;
 using Microsoft.CodeAnalysis.Text.Shared.Extensions;
 using Microsoft.VisualStudio.Text.Differencing;
@@ -38,13 +39,7 @@ internal class EditorTextDifferencingService(ITextDifferencingSelectorService di
 
         var differenceOptions = GetDifferenceOptions(preferredDifferenceType);
 
-        var oldTextSnapshot = oldText.FindCorrespondingEditorTextSnapshot();
-        var newTextSnapshot = newText.FindCorrespondingEditorTextSnapshot();
-        var useSnapshots = oldTextSnapshot != null && newTextSnapshot != null;
-
-        var diffResult = useSnapshots
-            ? diffService.DiffSnapshotSpans(oldTextSnapshot.GetFullSpan(), newTextSnapshot.GetFullSpan(), differenceOptions)
-            : diffService.DiffStrings(oldText.ToString(), newText.ToString(), differenceOptions);
+        var diffResult = diffService.DiffSourceTexts(oldText, newText, differenceOptions);
 
         return diffResult.Differences.Select(d =>
             new TextChange(
